feat: update script (#552)
