+++ scripts/fn/sleep.ts
@@ -1,0 +1,4 @@
+export default function sleep(time: number) {
+  console.log(`start sleep: ${time} ms`);
+  return new Promise((resolve) => setTimeout(resolve, time));
+}
